added agents differentiation message
