+++ src/utils/constants.py
@@ -1,0 +1,5 @@
+# Define the words for message generation
+WORDS = ["hello", "sun", "world", "space", "moon", "crypto", "sky", "ocean", "universe", "human"]
+
+# Accepted word types in the message
+TYPES_ACCEPTED = ['hello', 'crypto']
