+++ tests/__init__.py
@@ -1,0 +1,7 @@
+import os
+import sys
+PROJECT_PATH = os.getcwd()
+SOURCE_PATH = os.path.join(
+    PROJECT_PATH,"src"
+)
+sys.path.append(SOURCE_PATH)
